# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "AFSE API",
    "description": "API for Managing AFSE",
    "contact": {
      "name": "API Support",
      "email": "matteo.martinelli5@studenti.unimi.it"
    }
  },
  "host": "localhost:3000",
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Endpoint per registrare un nuovo utente",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "nome": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Nome dell utente"
                        },
                        "example": {
                          "type": "string",
                          "example": "Mario"
                        }
                      }
                    },
                    "cognome": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Cognome dell utente"
                        },
                        "example": {
                          "type": "string",
                          "example": "Rossi"
                        }
                      }
                    },
                    "username": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Username dell utente"
                        },
                        "example": {
                          "type": "string",
                          "example": "mrossi"
                        }
                      }
                    },
                    "email": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Email dell utente"
                        },
                        "example": {
                          "type": "string",
                          "example": "mario.rossi@example.com"
                        }
                      }
                    },
                    "password": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Password dell utente"
                        },
                        "example": {
                          "type": "string",
                          "example": "password123"
                        }
                      }
                    },
                    "profilePicture": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "URL dell immagine del profilo"
                        },
                        "example": {
                          "type": "string",
                          "example": "https://example.com/profile/mrossi.png"
                        }
                      }
                    },
                    "favouriteHero": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "integer"
                        },
                        "description": {
                          "type": "string",
                          "example": "ID del supereroe preferito dell utente"
                        },
                        "example": {
                          "type": "number",
                          "example": 1
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok"
          },
          "400": {
            "description": "Bad request"
          },
          "409": {
            "description": "Email/Username already exist."
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Effettua il login",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "Login credentials",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {
                          "type": "string",
                          "example": "user@example.com"
                        }
                      }
                    },
                    "password": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "example": {
                          "type": "string",
                          "example": "password123"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login ok"
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Invalid username or password"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Effettua il logout dell utente",
        "description": "Effettua il logout dell utente autenticato e invalida il JWT token. Richiede un JWT token passato tramite cookie.",
        "responses": {
          "200": {
            "description": "Logout effettuato con successo",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Messaggio di conferma"
                        },
                        "example": {
                          "type": "string",
                          "example": "Successfully logged out"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Utente non autenticato o JWT token mancante"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/auth/deleteAccount": {
      "delete": {
        "tags": [
          "Auth"
        ],
        "summary": "Elimina l account dell utente",
        "description": "Elimina l account dell utente autenticato. L ID dell utente è preso dal JWT token passato tramite cookie.",
        "responses": {
          "200": {
            "description": "Account dell utente eliminato con successo",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Messaggio di conferma"
                        },
                        "example": {
                          "type": "string",
                          "example": "User deleted successfully"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Utente non autenticato o JWT token mancante"
          },
          "404": {
            "description": "Utente non trovato",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Messaggio di errore"
                        },
                        "example": {
                          "type": "string",
                          "example": "Not found"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Impossibile completare la richiesta",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Messaggio di errore"
                        },
                        "example": {
                          "type": "string",
                          "example": "Cannot perform request"
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "jwt",
        "in": "cookie",
        "name": "jwt",
        "description": "JWT token passato tramite cookie"
      }
    },
...
  }
}